+++ lib/views/invitation_selection.dart
@@ -6,7 +6,6 @@ import 'package:fluffychat/components/avatar.dart';
 import 'package:fluffychat/components/matrix.dart';
 import 'package:fluffychat/i18n/i18n.dart';
 import 'package:flutter/material.dart';
-import 'package:toast/toast.dart';
 
 import 'chat_list.dart';
 
@@ -52,10 +51,12 @@ class _InvitationSelectionState extends State<InvitationSelection> {
       widget.room.invite(id),
     );
     if (success != false) {
-      Toast.show(
-        I18n.of(context).contactHasBeenInvitedToTheGroup,
-        context,
-        duration: Toast.LENGTH_LONG,
+      Scaffold.of(context).showSnackBar(
+        SnackBar(
+          content: Text(
+            I18n.of(context).contactHasBeenInvitedToTheGroup,
+          ),
+        ),
       );
     }
   }