+++ lib/views/sign_up_password.dart
@@ -7,7 +7,6 @@ import 'package:fluffychat/i18n/i18n.dart';
 import 'package:fluffychat/utils/app_route.dart';
 import 'package:fluffychat/views/auth_web_view.dart';
 import 'package:flutter/material.dart';
-import 'package:toast/toast.dart';
 
 import 'chat_list.dart';
 
@@ -96,7 +95,13 @@ class _SignUpPasswordState extends State<SignUpPassword> {
     try {
       await matrix.client.setDisplayname(widget.displayname);
     } catch (exception) {
-      Toast.show(I18n.of(context).couldNotSetDisplayname, context, duration: 5);
+      Scaffold.of(context).showSnackBar(
+        SnackBar(
+          content: Text(
+            I18n.of(context).couldNotSetDisplayname,
+          ),
+        ),
+      );
     }
     if (widget.avatar != null) {
       try {
@@ -107,7 +112,13 @@ class _SignUpPasswordState extends State<SignUpPassword> {
           ),
         );
       } catch (exception) {
-        Toast.show(I18n.of(context).couldNotSetAvatar, context, duration: 5);
+        Scaffold.of(context).showSnackBar(
+          SnackBar(
+            content: Text(
+              I18n.of(context).couldNotSetAvatar,
+            ),
+          ),
+        );
       }
     }
     await Navigator.of(context).pushAndRemoveUntil(